+++ Test/Project.AllPlatforms.Utility/Get-ProcessOutput.ps1
@@ -5,7 +5,7 @@ MIT License
 This file is part of "Windows Firewall Ruleset" project
 Homepage: https://github.com/metablaster/WindowsFirewallRuleset
 
-Copyright (C) 2019, 2020 metablaster zebal@protonmail.ch
+Copyright (C) 2020 metablaster zebal@protonmail.ch
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -26,10 +26,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 #>
 
-#
-# Unit test for Get-OneDrive
-#
-#Requires -RunAsAdministrator
+# Initialization
 . $PSScriptRoot\..\..\Config\ProjectSettings.ps1
 New-Variable -Name ThisScript -Scope Private -Option Constant -Value (
 	$MyInvocation.MyCommand.Name -replace ".{4}$" )
@@ -40,7 +37,6 @@ Initialize-Project -Abort
 # Imports
 . $PSScriptRoot\ContextSetup.ps1
 Import-Module -Name Project.AllPlatforms.Logging
-Import-Module -Name Project.Windows.UserInfo
 
 # User prompt
 Update-Context $TestContext $ThisScript @Logs
@@ -48,27 +44,12 @@ if (!(Approve-Execute -Accept $Accept -Deny $Deny @Logs)) { exit }
 
 Enter-Test $ThisScript
 
-$UserGroup = "Users"
+Start-Test "Get-ProcessOutput"
+$Result = Get-ProcessOutput -FilePath "git.exe" -ArgumentList "status" -NoNewWindow -Wait 3000 @Logs
+$Result
 
-Start-Test "Get-GroupPrincipal $UserGroup"
-$Principals = Get-GroupPrincipal $UserGroup @Logs
-# TODO: see also @Get-UserSoftware IF YOU REMOVE Format-Table THE TEST WILL NOT WORK!!
-$Principals | Format-Table
-
-foreach ($Principal in $Principals)
-{
-	Start-Test "Get-OneDrive $($Principal.User)"
-	Get-OneDrive $Principal.User @Logs
-}
-
-foreach ($Principal in $Principals)
-{
-	Start-Test "Get-OneDrive $($Principal.User) | InstallLocation"
-	Get-OneDrive $Principal.User @Logs | Select-Object -ExpandProperty InstallLocation
-}
-
-Start-Test "Get-TypeName - $($Principals[0].User)"
-Get-OneDrive $Principals[0].User @Logs | Get-TypeName @Logs
+Start-Test "Get-TypeName"
+$Result | Get-TypeName @Logs
 
 Update-Log
 Exit-Test